added a sakila.sqlite, the MySQL sakila sample database in a single sqlite file. php script has been modified to attach the sqlite file AS sakila. This allows everything to work out of the box - no need to setup a separate schema.
Unfortunately, the php script is now SQLite specific. 
It can be fixed only if we rethink the way connections are created. 
Important, but something we can solve real quick atm - have to think how to best implement this.

For now, we just focus on making a better MQL implementation, and we focus on db specific issues later on.